requirements.txt
opencv-python
opencv-contrib-python
numpy
Pillow
requests
openai
pymupdf